sphere: Add a sphere implementation
This patch extends the project by addition of a representation of
spheres and materials. By decoupling the two we allow spheres to reuse
materials without storing several instances of them.

We also add a very simple and lacking raycasting implementation to see
if the result vaguely makes sense and use it instead of the random
pattern that was previously displayed in the window.

Signed-off-by: Dus'li <duszku511@gmail.com>