scene: Update render kernel to use light sources
I don't think this is proper Phong reflection model yet. The results
just look... ugly? I must have done something wrong.

Moreover, it is out of the question for us to support larger quantities
of spheres or lights, as everything is implemented super suboptimally.

But it is a step forward nonetheless.

Signed-off-by: Dus'li <duszku511@gmail.com>