refactor: Adjust constants, cleanup code, comment
This is a chunky patch that probably should be broken into pieces. It is
meant to refine the existing solution into something that actually looks
nice. It is also meant to turn the code into something that is
(relatively) comprehensible to read.

Key changes are:

  - guaranteeing that material constants sum up to 1,
  - limiting shininess to a reasonable interval,
  - tone mapping with Reinhard method,
  - gamma correction,
  - better light location randomization,
  - adjusted light source constants,
  - adding light falloff over distance.

The vecops.cuh was also heavily reworked so as not to repeat myself so
much and allow for easier extensions in the future. Not everybody likes
X macros, but I do and I found this argument convincing.

Signed-off-by: Dus'li <duszku511@gmail.com>