scene: Implement camera and camera movement
This one's a wee bit big. But it is structurally cohesive. Basically
we're adding a camera class, use it in the renderer, add callback
registration for SDL windows, then register handlers for mouse pointer
movement and wasd keys and use them to implement a poor FPS-like camera.
And then we wrap everything in appstate so that main() doesn't get too
cluttered.

Signed-off-by: Dus'li <duszku511@gmail.com>