sphere: Expose host-side representation
This patch unwraps the meticulously designed layer of abstraction that
is SphereSet and spawns a new struct: HostSphereSet. The on-host
initialization of sphere data thus became an explicit step in the
application boot process.

This is the first step in implementing an Octree to address the
performance issues observable for large quantities of spheres.

Signed-off-by: Dus'li <duszku511@gmail.com>